io-sim-por: Moved types to Control.Monad.IOSimPOR.Types
