+++ io-sim/src/Control/Monad/IOSim/Types.hs
@@ -1107,63 +1107,6 @@ data StmStack s b a where
                    -> StmStack s b c
                    -> StmStack s a c
 
-
----
---- Schedules
----
-
--- | Modified execution schedule.
---
-data ScheduleControl = ControlDefault
-                     -- ^ default scheduling mode
-                     | ControlAwait [ScheduleMod]
-                     -- ^ if the current control is 'ControlAwait', the normal
-                     -- scheduling will proceed, until the thread found in the
-                     -- first 'ScheduleMod' reaches the given step.  At this
-                     -- point the thread is put to sleep, until after all the
-                     -- steps are followed.
-                     | ControlFollow [StepId] [ScheduleMod]
-                     -- ^ follow the steps then continue with schedule
-                     -- modifications.  This control is set by 'followControl'
-                     -- when 'controlTargets' returns true.
-  deriving (Eq, Ord, Show)
-
-
-isDefaultSchedule :: ScheduleControl -> Bool
-isDefaultSchedule ControlDefault        = True
-isDefaultSchedule (ControlFollow [] []) = True
-isDefaultSchedule _                     = False
-
--- | A schedule modification inserted at given execution step.
---
-data ScheduleMod = ScheduleMod{
-    -- | Step at which the 'ScheduleMod' is activated.
-    scheduleModTarget    :: StepId,
-    -- | 'ScheduleControl' at the activation step.  It is needed by
-    -- 'extendScheduleControl' when combining the discovered schedule with the
-    -- initial one.
-    scheduleModControl   :: ScheduleControl,
-    -- | Series of steps which are executed at the target step.  This *includes*
-    -- the target step, not necessarily as the last step.
-    scheduleModInsertion :: [StepId]
-  }
-  deriving (Eq, Ord)
-
--- | Execution step is identified by the thread id and a monotonically
--- increasing number (thread specific).
---
-type StepId = (ThreadId, Int)
-
-instance Show ScheduleMod where
-  showsPrec d (ScheduleMod tgt ctrl insertion) =
-    showParen (d>10) $
-      showString "ScheduleMod " .
-      showsPrec 11 tgt .
-      showString " " .
-      showsPrec 11 ctrl .
-      showString " " .
-      showsPrec 11 insertion
-
 ---
 --- Exploration options
 ---
